Break down transaction into components
